requirements1.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

tensorflow==2.13.1

flask>=2.2.0
Pillow>=9.0.0
numpy>=1.21
tqdm>=4.64.0

# Optionnel pour debug et test
matplotlib>=3.5.0

